refactor: break the 8-package import cycle (Phase 0+1)
Collapse the strongly-connected component of
{archive, backend, benchmark, cli, git, model, remote, versioning}
down to two isolated 2-loops by relocating shared utilities out of the
cli layer and dissolving the utils catch-alls.

Phase 0 — dissolve catch-alls:
  - merge_dict_list  -> model/_merge.py
  - format_mc_output -> benchmark/_mc_output.py
  - delete test-only dead code: as_list, try_avail_envmodule,
    parse_extra_args (+ their tests)
  - remove omnibenchmark/utils.py and cli/utils/

Phase 1 — lift utilities out of cli:
  - cli/utils/logging.py    -> omnibenchmark/logging.py
  - format_yaml_warning     -> omnibenchmark/error_formatting.py
    (pretty_print_parse_error stays in cli; it depends on model)
  - cli/progress.py         -> omnibenchmark/progress.py

Nothing imports back up into cli anymore. Cycle edges 10 -> 4.
Remaining loops (benchmark<->git, archive<->remote) are Phases 2-3.
