+++ tests/test_utils.py
@@ -6,163 +6,11 @@ All tests are marked as 'short' since they test simple utility functions.
 """
 
 import pytest
-import os
 from pathlib import Path
-from unittest.mock import Mock, patch
+from unittest.mock import Mock
 
-from omnibenchmark.utils import (
-    try_avail_envmodule,
-    as_list,
-    merge_dict_list,
-    format_mc_output,
-)
-
-
-@pytest.mark.short
-class TestTryAvailEnvmodule:
-    """Test the try_avail_envmodule function."""
-
-    @patch("subprocess.run")
-    def test_module_available_returns_true(self, mock_run):
-        """Test that available module returns True."""
-        # Mock successful module check (no "No module" message in stderr)
-        mock_run.return_value = Mock(
-            stderr="Available Modules:\n  gcc/11.2.0\n  gcc/12.1.0"
-        )
-
-        result = try_avail_envmodule("gcc")
-        assert result is True
-
-        # Verify the command structure
-        mock_run.assert_called_once()
-        call_args = mock_run.call_args
-        assert "module avail gcc" in call_args[0][0]
-        assert call_args[1]["shell"] is True
-        assert call_args[1]["text"] is True
-
-    @patch("subprocess.run")
-    def test_module_unavailable_returns_false(self, mock_run):
-        """Test that unavailable module returns False."""
-        # Mock failed module check
-        mock_run.return_value = Mock(stderr="No module(s) or extension(s) found!")
-
-        result = try_avail_envmodule("nonexistent-module")
-        assert result is False
-
-    @patch("subprocess.run")
-    def test_partial_match_in_error_returns_false(self, mock_run):
-        """Test that partial error message match still returns False."""
-        mock_run.return_value = Mock(
-            stderr="Some other error. No module(s) or extension(s) found! More text."
-        )
-
-        result = try_avail_envmodule("test-module")
-        assert result is False
-
-    @patch("subprocess.run")
-    def test_empty_stderr_returns_true(self, mock_run):
-        """Test that empty stderr (successful check) returns True."""
-        mock_run.return_value = Mock(stderr="")
-
-        result = try_avail_envmodule("some-module")
-        assert result is True
-
-    @patch("subprocess.run")
-    def test_environment_variables_passed(self, mock_run):
-        """Test that environment variables are properly passed to subprocess."""
-        mock_run.return_value = Mock(stderr="")
-
-        # Set a test environment variable
-        test_env_var = "TEST_OMNIBENCH_VAR"
-        test_env_value = "test_value"
-        os.environ[test_env_var] = test_env_value
-
-        try:
-            try_avail_envmodule("test-module")
-
-            # Check that environment was passed
-            call_args = mock_run.call_args
-            passed_env = call_args[1]["env"]
-            assert test_env_var in passed_env
-            assert passed_env[test_env_var] == test_env_value
-        finally:
-            # Clean up
-            if test_env_var in os.environ:
-                del os.environ[test_env_var]
-
-    @patch("subprocess.run")
-    def test_command_structure(self, mock_run):
-        """Test that the command has the expected structure."""
-        mock_run.return_value = Mock(stderr="")
-
-        try_avail_envmodule("test-module")
-
-        call_args = mock_run.call_args
-        command = call_args[0][0]
-
-        # Verify command components
-        assert '. "$LMOD_PKG"/init/profile' in command
-        assert "module purge" in command
-        assert "module avail test-module" in command
-
-
-@pytest.mark.short
-class TestAsList:
-    """Test the as_list utility function."""
-
-    def test_list_input_returns_same_list(self):
-        """Test that list input returns the same list."""
-        input_list = [1, 2, 3, "test"]
-        result = as_list(input_list)
-        assert result == input_list
-        assert result is input_list  # Should be the same object
-
-    def test_string_input_returns_list_with_string(self):
-        """Test that string input returns list containing the string."""
-        input_str = "test_string"
-        result = as_list(input_str)
-        assert result == ["test_string"]
-        assert isinstance(result, list)
-        assert len(result) == 1
-
-    def test_integer_input_returns_list_with_integer(self):
-        """Test that integer input returns list containing the integer."""
-        input_int = 42
-        result = as_list(input_int)
-        assert result == [42]
-        assert isinstance(result, list)
-        assert len(result) == 1
-
-    def test_none_input_returns_list_with_none(self):
-        """Test that None input returns list containing None."""
-        result = as_list(None)
-        assert result == [None]
-        assert isinstance(result, list)
-        assert len(result) == 1
-
-    def test_dict_input_returns_list_with_dict(self):
-        """Test that dict input returns list containing the dict."""
-        input_dict = {"key": "value", "number": 123}
-        result = as_list(input_dict)
-        assert result == [input_dict]
-        assert isinstance(result, list)
-        assert len(result) == 1
-        assert result[0] is input_dict
-
-    def test_empty_list_returns_empty_list(self):
-        """Test that empty list returns empty list."""
-        input_list = []
-        result = as_list(input_list)
-        assert result == []
-        assert result is input_list
-
-    def test_tuple_input_returns_list_with_tuple(self):
-        """Test that tuple input returns list containing the tuple."""
-        input_tuple = (1, 2, 3)
-        result = as_list(input_tuple)
-        assert result == [(1, 2, 3)]
-        assert isinstance(result, list)
-        assert len(result) == 1
+from omnibenchmark.model._merge import merge_dict_list
+from omnibenchmark.benchmark._mc_output import format_mc_output
 
 
 @pytest.mark.short
